Housekeeping (#40)
**Any of these commits may be dropped or rebased should you
disapprove.**

I recommend reviewing this PR by individual commit because of the
diff-noise produced by the formatter (last commit).

1. Add `meta.mainProgram` (and other `meta` attributes) to the package.

<https://github.com/MercuryTechnologies/nix-your-shell/pull/40/commits/af3c6a8820135c0897e77b7cc6666e99c78a3630>
3. Appease `nil` lint about unused function argument `flake-compat`,
replace with `...`.

<https://github.com/MercuryTechnologies/nix-your-shell/pull/40/commits/079602288895d5a1ebb7bbdd670cad1611235a98>
5. Remove usage of `rec` from `packages` output, prefer referencing
through `self` for package aliases.

<https://github.com/MercuryTechnologies/nix-your-shell/pull/40/commits/30ebc4953280dcced19fdab8f1ad94a688cfd3bf>
7. Remove `system` from `import nixpkgs` arguments, prefer
`localSystem`.

<https://github.com/MercuryTechnologies/nix-your-shell/pull/40/commits/82a5d248c3fe0cd1bf06666af2324f7d5b8ba12d>
   - This is a documented change in Nixpkgs, but a relatively quiet one.
- The original `system` argument has been referred to as "legacy" since
2017.
- See the source:
<https://github.com/NixOS/nixpkgs/blob/5d2b2523a0442fd19294359b31b8535696c32984/pkgs/top-level/impure.nix#L14>
- See the commit that changed this:
<https://github.com/NixOS/nixpkgs/commit/8cd4c31d6b8c6432986edf0ec5d00d2fe1c12854>
- Reference to `stdenv`:
<https://github.com/NixOS/nixpkgs/blob/5d2b2523a0442fd19294359b31b8535696c32984/pkgs/stdenv/default.nix#L10>
8. Remove `flake-utils`.

<https://github.com/MercuryTechnologies/nix-your-shell/pull/40/commits/61d2ab423753e4ebc799cd39ba7a4b80cf9bc817>
   - See: <https://ayats.org/blog/no-flake-utils/>
- The title of the latest release for `flake-utils` from 2022 is: "It's
a stable mess."
9. Introduce [nix-systems](https://github.com/nix-systems/nix-systems)
to replace functionality of `flake-utils`.

<https://github.com/MercuryTechnologies/nix-your-shell/pull/40/commits/898524afa011e8eb2b6f431406a9afbf8c8e9b2d>
- This makes `systems` overrideable through usage of
`nix-your-shell.inputs.systems`.
10. Adopt `nixfmt` as the flake's `formatter`.

<https://github.com/MercuryTechnologies/nix-your-shell/pull/40/commits/d47c7a1ee367fb82805d44fc1a92e94459f97bed>
- This is perhaps somewhat controversial. There was discussion in the
issues of `nixfmt` with members of the NixOS organization about
"blessing" this as the canon formatter. Unfortunately, contrary to the
style preferred by many individual flake projects, they decided not to
go through with it and instead are "waiting" for `nixpkgs-fmt` to be
"ironed out".
    - I chose this one because:
      1. I like it.
      2. It matches the style you were already using, mostly.
- I had to change the `sed` script and the `LOAD-BEARING COMMENT` to
make it shorter, so that `nixfmt` doesn't wrap the line. Since `sed`
can't do multiline matching replacement easily, this was the solution
that resisted me the least. I recommend taking a look at its
replacement, and advising me further if need-be. No matter the flake's
`formatter`, making this mechanism less fragine would be a good idea.
     
I will also look into making another PR down the line (once I figure it
out for myself) about simplifying the `version.yaml` workflow. This
makes me uneasy, having a `sed` script target a comment. Ideally, that
version should be based on the flake's revision, however I do understand
it is necessary to be the way it is because of publishing to
[crates.io](https://crates.io).